% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_sequence_cos.R
\name{DNA_sequence_cos}
\alias{DNA_sequence_cos}
\title{Obtain all mutations codons}
\usage{
DNA_sequence_cos(cosmic, num = length(cosmic$GENE_NAME), random = FALSE)
}
\arguments{
\item{cosmic}{Dataframe output from the import_data_cos function}

\item{num}{The number of patient you want to analyse. Default will go through all the dataset.}
}
\value{
DNA_seq = CDS sequence for all the genes used, codon_part = codon around mutation.
}
\description{
This function will load all the CDS from the mutated genes in the input file. Then recovery the 3' and
5' nucleotide aside of the mutation site.
}
