% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_cos.R
\name{import_data_cos}
\alias{import_data_cos}
\title{Load a Cosmic targeted screens mutation data}
\usage{
import_data_cos(file)
}
\arguments{
\item{file}{Path to the input file}
}
\value{
A dataframe with all usable patient
}
\description{
This function will load and delete all the no needed patient with NA values or Null characters.
}
