% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_signature_cos.R
\name{mut_signature_cos}
\alias{mut_signature_cos}
\title{Create the final results Dataframe}
\usage{
mut_signature_cos(mut_to, file = 1)
}
\arguments{
\item{mut_to}{Codon of mutation paste to the change of nucleotide due to mutation.}

\item{file}{Just the number of the file analyzed. Default = 1}
}
\value{
mut_signa = dataset with the counts of each mutations
}
\description{
This function will create a dataframe with all possible mutation and count and calculate the frequency
of each from the input dataset.
}
