% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotbar.R
\name{Plotbar}
\alias{Plotbar}
\title{Create mutational signature from a cosmic dataset}
\usage{
Plotbar(total)
}
\arguments{
\item{total}{Path to the output of Signature_algoV2 function.}
}
\value{
Mutational signature using ggplot2.
}
\description{
This function will create mutational signature from an input dataset. See each function description
for more details.
}
