% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_mut_cos.R
\name{Site_mut_cos}
\alias{Site_mut_cos}
\title{Substitute same molecular mutation}
\usage{
Site_mut_cos(seq)
}
\arguments{
\item{seq}{Output from DNA_sequence_cos function}
}
\value{
change_nu = the change of nucleotide due to the mutation, mut_to = the codon paste to
the nucleotide after mutation.
}
\description{
This function will just change nucleotide from mutation to get only T and C mutated.
}
